add commands creating users
