+++ mailing/views.py
@@ -194,7 +194,8 @@ class TransmissionView(ListView):
     def get_context_data(self, *, object_list=None, **kwargs):
         context = super().get_context_data(**kwargs)
         context["Title"] = "Transmissions"
-        context["Transmissions"] = Transmission.objects.filter(owner=self.request.user)
+        # context["Transmissions"] = Transmission.objects.filter(owner=self.request.user)
+        context["Transmissions"] = Transmission.objects.all()
         return context
