--- users/management/commands/create_users.py
@@ -1,29 +1,0 @@
-from django.core.management import BaseCommand
-from users.models import User
-
-
-class Command(BaseCommand):
-
-    def handle(self, *args, **options):
-        user = User.objects.create(
-            email="admin@gmail.com",
-            first_name="admin",
-            last_name="admin",
-            is_superuser=True,
-            is_staff=True,
-            is_active=True
-            )
-
-        user.set_password("admin")
-        user.save()
-
-        user = User.objects.create(
-            email="moderator@gmail.com",
-            first_name="moderator",
-            last_name="moderator",
-            is_superuser=False,
-            is_staff=True,
-            is_active=True
-        )
-        user.set_password("moderator")
-        user.save()
